refactor: update a components
