+++ src/components/header/GlobalHeader.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  export let title: string
+</script>
+
+<svelte:head>
+  <title>{title}</title>
+</svelte:head>
