+++ src/components/header/Header.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  export let title: string
+
+  import './Header.scss'
+</script>
+
+<div class="container">
+  <h1 class="container-header">{title}</h1>
+</div>
