+++ src/components/image/Image.scss
@@ -1,0 +1,11 @@
+figure {
+  margin: 0 0 1em 0;
+  text-align: center;
+  margin: 0 auto;
+}
+
+img {
+  width: 100%;
+  max-width: 400px;
+  margin: 0 0 1em 0;
+}
