+++ src/components/navbar/Nav.scss
@@ -1,0 +1,35 @@
+nav {
+  border-bottom: 1px solid rgba(255, 62, 0, 0.1);
+  font-weight: 300;
+  padding: 0 1em;
+}
+
+ul {
+  margin: 0;
+  padding: 0;
+  display: flex;
+  align-items: center;
+}
+
+/* clearfix */
+
+ul::after {
+  content: '';
+  clear: both;
+}
+
+.selected {
+  position: relative;
+  display: flex;
+  align-items: center;
+}
+
+.selected::after {
+  position: absolute;
+  content: '';
+  width: calc(100% - 1em);
+  height: 2px;
+  background-color: rgb(255, 62, 0);
+  display: block;
+  bottom: -1px;
+}
