+++ src/components/navbar/Nav.utils.ts
@@ -1,0 +1,5 @@
+import { logger } from '../../helpers/logger.helper'
+
+export const showSegment = (segment: string): void => {
+  logger(segment)
+}
