+++ src/components/navbar/NavLink.scss
@@ -1,0 +1,27 @@
+.selected {
+  position: relative;
+  display: flex;
+  align-items: center;
+
+  &::after {
+    position: absolute;
+    content: '';
+    width: calc(100% - 1em);
+    height: 2px;
+    background-color: rgb(255, 62, 0);
+    display: block;
+    bottom: -1px;
+  }
+}
+
+a {
+  text-decoration: none;
+  padding: 1em 0.5em;
+  display: flex;
+  align-items: center;
+}
+
+li {
+  display: block;
+  float: left;
+}
