+++ src/components/title/Title.scss
@@ -1,0 +1,4 @@
+p {
+  margin: 1em auto;
+  text-align: center;
+}
