+++ src/components/title/Title.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  export let text: string
+
+  import './Title.scss'
+</script>
+
+<p>
+  <strong>{text}</strong>
+</p>
