+++ src/helpers/logger.helper.ts
@@ -1,3 +1,7 @@
 export const logger = (args: unknown): void => {
   console.log(args)
 }
+
+export const loggerWithDate = (args: unknown): void => {
+  console.log(args, new Date())
+}
