+++ src/routes/_layout.svelte
@@ -1,32 +1,11 @@
 <script lang="ts">
-  import Nav from '../components/Nav.svelte'
+  import Nav from '../components/navbar/Nav.svelte'
 
-  import type ITab from '../interfaces/tab.interface'
+  import { tabs } from '../helpers/tab.helper'
 
   import './layout.scss'
 
   export let segment: string
-
-  const tabs: ITab[] = [
-    {
-      segment: undefined,
-      href: '.',
-      title: 'home',
-      prefetch: false,
-    },
-    {
-      segment: 'about',
-      href: 'about',
-      title: 'about',
-      prefetch: false,
-    },
-    {
-      segment: 'blog',
-      href: 'blog',
-      title: 'blog',
-      prefetch: true,
-    },
-  ]
 </script>
 
 <Nav {segment} {tabs} />
