--- src/routes/index.scss
@@ -1,12 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-
-  &-header {
-    font-size: 42px;
-    display: flex;
-    justify-content: center;
-    width: 100%;
-  }
-}
