+++ src/routes/index.svelte
@@ -1,58 +1,20 @@
-<style>
-  h1,
-  figure,
-  p {
-    text-align: center;
-    margin: 0 auto;
-  }
+<script lang="ts">
+  import GlobalHeader from '../components/header/GlobalHeader.svelte'
+  import Header from '../components/header/Header.svelte'
+  import Image from '../components/image/Image.svelte'
+  import Title from '../components/title/Title.svelte'
 
-  h1 {
-    font-size: 2.8em;
-    text-transform: uppercase;
-    font-weight: 700;
-    margin: 0 0 0.5em 0;
-  }
+  import { loggerWithDate } from '../helpers/logger.helper'
 
-  figure {
-    margin: 0 0 1em 0;
-  }
-
-  img {
-    width: 100%;
-    max-width: 400px;
-    margin: 0 0 1em 0;
-  }
-
-  p {
-    margin: 1em auto;
-  }
-
-  @media (min-width: 480px) {
-    h1 {
-      font-size: 4em;
-    }
-  }
-</style>
-
-<script>
   import './index.scss'
 
-  console.log('Started:', new Date())
+  loggerWithDate('Started:')
 </script>
 
-<svelte:head>
-  <title>Sapper project template</title>
-</svelte:head>
+<GlobalHeader title="Sapper project template" />
 
-<div class="container">
-  <h1 class="container-header">Great success!</h1>
-</div>
+<Header title="{'Great success!'}" />
 
-<figure>
-  <img alt="Borat" src="assets/img/great-success.png" />
-  <figcaption>HIGH FIVE!</figcaption>
-</figure>
+<Image caption="{'HIGH FIVE!'}" alt="borat" src="assets/img/great-success.png" />
 
-<p>
-  <strong>Try editing this file (src/routes/index.svelte) to test live reloading.</strong>
-</p>
+<Title text="Try editing this file (src/routes/index.svelte) to test live reloading." />
